+++ netket/models/equivariant.py
@@ -70,8 +70,6 @@ class GCNN_FFT(nn.Module):
     all layers will have the same number of features."""
     characters: HashableArray
     """Array specifying the characters of the desired symmetry representation"""
-    in_features: int = 1
-    """Number of input features, defaults to 1."""
     dtype: Any = float
     """The dtype of the weights."""
     activation: Any = reim_selu
@@ -96,8 +94,7 @@ class GCNN_FFT(nn.Module):
         self.dense_symm = DenseSymmFFT(
             space_group=self.symmetries,
             shape=self.shape,
-            out_features=self.features[0],
-            in_features=self.in_features,
+            features=self.features[0],
             dtype=self.dtype,
             use_bias=self.use_bias,
             kernel_init=self.kernel_init,
@@ -109,7 +106,6 @@ class GCNN_FFT(nn.Module):
             DenseEquivariantFFT(
                 product_table=self.product_table,
                 shape=self.shape,
-                in_features=self.features[layer],
                 out_features=self.features[layer + 1],
                 use_bias=self.use_bias,
                 dtype=self.dtype,
@@ -180,8 +176,6 @@ class GCNN_Irrep(nn.Module):
     all layers will have the same number of features."""
     characters: HashableArray
     """Array specifying the characters of the desired symmetry representation"""
-    in_features: int = 1
-    """Number of input features, defaults to 1."""
     dtype: Any = np.float64
     """The dtype of the weights."""
     activation: Any = reim_selu
@@ -205,8 +199,7 @@ class GCNN_Irrep(nn.Module):
 
         self.dense_symm = DenseSymmMatrix(
             symmetries=self.symmetries,
-            out_features=self.features[0],
-            in_features=self.in_features,
+            features=self.features[0],
             dtype=self.dtype,
             use_bias=self.use_bias,
             kernel_init=self.kernel_init,
@@ -217,7 +210,6 @@ class GCNN_Irrep(nn.Module):
         self.equivariant_layers = [
             DenseEquivariantIrrep(
                 irreps=self.irreps,
-                in_features=self.features[layer],
                 out_features=self.features[layer + 1],
                 use_bias=self.use_bias,
                 dtype=self.dtype,
@@ -275,8 +267,6 @@ class GCNN_Parity_FFT(nn.Module):
     """Array specifying the characters of the desired symmetry representation"""
     parity: int
     """Integer specifying the eigenvalue with respect to parity"""
-    in_features: int = 1
-    """Number of input features, defaults to 1."""
     dtype: Any = np.float64
     """The dtype of the weights."""
     activation: Any = reim_selu
@@ -301,8 +291,7 @@ class GCNN_Parity_FFT(nn.Module):
         self.dense_symm = DenseSymmFFT(
             space_group=self.symmetries,
             shape=self.shape,
-            out_features=self.features[0],
-            in_features=self.in_features,
+            features=self.features[0],
             dtype=self.dtype,
             use_bias=self.use_bias,
             kernel_init=self.kernel_init,
@@ -314,7 +303,6 @@ class GCNN_Parity_FFT(nn.Module):
             DenseEquivariantFFT(
                 product_table=self.product_table,
                 shape=self.shape,
-                in_features=self.features[layer],
                 out_features=self.features[layer + 1],
                 use_bias=self.use_bias,
                 dtype=self.dtype,
@@ -329,7 +317,6 @@ class GCNN_Parity_FFT(nn.Module):
             DenseEquivariantFFT(
                 product_table=self.product_table,
                 shape=self.shape,
-                in_features=self.features[layer],
                 out_features=self.features[layer + 1],
                 # this would bias the same outputs as self.equivariant
                 use_bias=False,
@@ -433,8 +420,6 @@ class GCNN_Parity_Irrep(nn.Module):
     """Array specifying the characters of the desired symmetry representation"""
     parity: int
     """Integer specifying the eigenvalue with respect to parity"""
-    in_features: int = 1
-    """Number of input features, defaults to 1."""
     dtype: Any = np.float64
     """The dtype of the weights."""
     activation: Any = reim_selu
@@ -458,8 +443,7 @@ class GCNN_Parity_Irrep(nn.Module):
 
         self.dense_symm = DenseSymmMatrix(
             symmetries=self.symmetries,
-            out_features=self.features[0],
-            in_features=self.in_features,
+            features=self.features[0],
             dtype=self.dtype,
             use_bias=self.use_bias,
             kernel_init=self.kernel_init,
@@ -470,7 +454,6 @@ class GCNN_Parity_Irrep(nn.Module):
         self.equivariant_layers = [
             DenseEquivariantIrrep(
                 irreps=self.irreps,
-                in_features=self.features[layer],
                 out_features=self.features[layer + 1],
                 use_bias=self.use_bias,
                 dtype=self.dtype,
@@ -484,7 +467,6 @@ class GCNN_Parity_Irrep(nn.Module):
         self.equivariant_layers_flip = [
             DenseEquivariantIrrep(
                 irreps=self.irreps,
-                in_features=self.features[layer],
                 out_features=self.features[layer + 1],
                 # this would bias the same outputs as self.equivariant
                 use_bias=False,
@@ -589,7 +571,6 @@ def GCNN(
         layers: Number of layers (not including sum layer over output).
         features: Number of features in each layer starting from the input. If a single
             number is given, all layers will have the same number of features.
-        in_features: Number of input features, defaults to 1.
         characters: Array specifying the characters of the desired symmetry representation
         parity: Optional argument with value +/-1 that specifies the eigenvalue
             with respect to parity (only use on two level systems).